requirements.txt
python-dotenv==1.0.1
pandas==2.2.3
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
openpyxl==3.1.5
